Trying to separate style from components
