+++ src/style/NavbarStyle.tsx
@@ -1,0 +1,15 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme) => ({
+	root: {
+	  flexGrow: 1,
+	},
+	menuButton: {
+	  marginRight: theme.spacing(2),
+	},
+	title: {
+	  flexGrow: 1,
+	},
+}));  
+
+export default { useStyles };
